Create 1. Santa's Present Factory
Christmas is just around the corner and this year Santa has decided to share his secret with you. Get ready to learn how his dwarfs craft all the presents. 
First you will receive a sequence of integers, representing the number of materials for crafting toys in one box. After that you will be given another sequence of integers – their magic level.
Your task is to mix materials with magic so you can craft presents, listed in the table below with the exact magic level. 
In order to craft a toy, you have to take the last box with materials and the first magic level value. The total magic level is calculated by their multiplication. If the result equals one of the levels described in the table above, you craft the present and remove both materials and magic value. Otherwise:
•	If the product of the operation is a negative number, then you have to sum the values together, remove them both from their positions and the result should be added to the materials.
•	If the product doesn’t equal one of the magic levels in the table and is a positive number, remove only the magic value and increase the material value with 15.
•	If the magic or material (or both) equals 0, remove it (or both) and continue crafting the presents.
Stop crafting presents when you run out of boxes of materials or magic level values.
Your task is considered done if you manage to craft either one of the pairs - a doll and a train or a teddy bear and a bicycle.
Input
•	The first line of input will represent the values of boxes with materials - integers, separated by single space
•	On the second line you will be given the magic values - integers again, separated by single space
Output
•	On the first line - print whether you've succeeded in crafting the presents
o	"The presents are crafted! Merry Christmas!"
o	"No presents this Christmas!"
•	On the next two lines print the materials and magic that are left, if there are any, otherwise skip the line
o	"Materials left: {material1}, {material2}, …"
o	"Magic left: {magicValue1}, {magicValue2}, …
•	On the next lines print the presents you have crafted at least once, ordered alphabetically in the format:
"{toy name}: {amount}"
Constraints
•	All of the materials' values will be integers in range [1, 100]
•	Magic level values will be integers in range [-10, 100]
•	In all cases at least one present will be crafted
// 10 -5 20 15 -30 10
    40 60 10 4 10 0

//30 5 15 60 0 30
  -15 10 5 -15 25